// apps/api-videos/src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  Json: { input: any; output: any; }
  join__DirectiveArguments: { input: any; output: any; }
  join__FieldSet: { input: any; output: any; }
  link__Import: { input: any; output: any; }
};

export type Action = {
  gtmEventName?: Maybe<Scalars['String']['output']>;
  parentBlock: Block;
  parentBlockId: Scalars['ID']['output'];
};

export type AudioPreview = {
  __typename?: 'AudioPreview';
  duration: Scalars['Int']['output'];
  language: Language;
  size: Scalars['Int']['output'];
  value: Scalars['String']['output'];
};

export type Block = {
  id: Scalars['ID']['output'];
  journeyId: Scalars['ID']['output'];
  parentBlockId?: Maybe<Scalars['ID']['output']>;
  parentOrder?: Maybe<Scalars['Int']['output']>;
};
// ...
export type CloudflareImage = {
  __typename?: 'CloudflareImage';
  createdAt: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  uploadUrl?: Maybe<Scalars['String']['output']>;
  userId: Scalars['ID']['output'];
};

export type CloudflareVideo = {
  __typename?: 'CloudflareVideo';
  createdAt: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  readyToStream: Scalars['Boolean']['output'];
  uploadUrl?: Maybe<Scalars['String']['output']>;
  userId: Scalars['ID']['output'];
};

export type Country = {
  __typename?: 'Country';
  continent: Array<CountryContinent>;
  flagPngSrc?: Maybe<Scalars['String']['output']>;
  flagWebpSrc?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  languages: Array<Language>;
  latitude?: Maybe<Scalars['Float']['output']>;
  longitude?: Maybe<Scalars['Float']['output']>;
  name: Array<CountryName>;
  population?: Maybe<Scalars['Int']['output']>;
};


export type CountryContinentArgs = {
  languageId?: InputMaybe<Scalars['ID']['input']>;
  primary?: InputMaybe<Scalars['Boolean']['input']>;
};


export type CountryNameArgs = {
  languageId?: InputMaybe<Scalars['ID']['input']>;
  primary?: InputMaybe<Scalars['Boolean']['input']>;
};

export type CountryContinent = {
  __typename?: 'CountryContinent';
  language: Language;
  primary: Scalars['Boolean']['output'];
  value: Scalars['String']['output'];
};

export type CountryName = {
  __typename?: 'CountryName';
  language: Language;
  primary: Scalars['Boolean']['output'];
  value: Scalars['String']['output'];
};

export type CreateVerificationRequestInput = {
  redirect?: InputMaybe<Scalars['String']['input']>;
};

export type CustomDomain = {
  __typename?: 'CustomDomain';
  apexName: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  journeyCollection?: Maybe<JourneyCollection>;
  name: Scalars['String']['output'];
  routeAllTeamJourneys: Scalars['Boolean']['output'];
  team: Team;
};
// ...
export type Host = {
  __typename?: 'Host';
  id: Scalars['ID']['output'];
  location?: Maybe<Scalars['String']['output']>;
  src1?: Maybe<Scalars['String']['output']>;
  src2?: Maybe<Scalars['String']['output']>;
  teamId: Scalars['ID']['output'];
  title: Scalars['String']['output'];
};
// ...
export type Integration = {
  id: Scalars['ID']['output'];
  team: Team;
  type: IntegrationType;
};
// ...
export type Journey = {
  __typename?: 'Journey';
  archivedAt?: Maybe<Scalars['DateTime']['output']>;
  blocks?: Maybe<Array<Block>>;
  chatButtons: Array<ChatButton>;
  createdAt: Scalars['DateTime']['output'];
  creatorDescription?: Maybe<Scalars['String']['output']>;
  creatorImageBlock?: Maybe<ImageBlock>;
  deletedAt?: Maybe<Scalars['DateTime']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  featuredAt?: Maybe<Scalars['DateTime']['output']>;
  host?: Maybe<Host>;
  id: Scalars['ID']['output'];
  journeyCollections: Array<JourneyCollection>;
  language: Language;
  /** used in a plausible share link to embed report */
  plausibleToken?: Maybe<Scalars['String']['output']>;
  primaryImageBlock?: Maybe<ImageBlock>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  seoDescription?: Maybe<Scalars['String']['output']>;
  seoTitle?: Maybe<Scalars['String']['output']>;
  slug: Scalars['String']['output'];
  status: JourneyStatus;
  strategySlug?: Maybe<Scalars['String']['output']>;
  tags: Array<Tag>;
  team?: Maybe<Team>;
  template?: Maybe<Scalars['Boolean']['output']>;
  themeMode: ThemeMode;
  themeName: ThemeName;
  title: Scalars['String']['output'];
  trashedAt?: Maybe<Scalars['DateTime']['output']>;
  userJourneys?: Maybe<Array<UserJourney>>;
};

export type JourneyCollection = {
  __typename?: 'JourneyCollection';
  customDomains?: Maybe<Array<CustomDomain>>;
  id: Scalars['ID']['output'];
  journeys?: Maybe<Array<Journey>>;
  team: Team;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type JourneyProfile = {
  __typename?: 'JourneyProfile';
  acceptedTermsAt?: Maybe<Scalars['DateTime']['output']>;
  id: Scalars['ID']['output'];
  journeyFlowBackButtonClicked?: Maybe<Scalars['Boolean']['output']>;
  lastActiveTeamId?: Maybe<Scalars['String']['output']>;
  onboardingFormCompletedAt?: Maybe<Scalars['DateTime']['output']>;
  userId: Scalars['ID']['output'];
};
// ...
export type JourneyVisitorsConnection = {
  __typename?: 'JourneyVisitorsConnection';
  /** A list of edges. */
  edges: Array<JourneyVisitorEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

export type JourneysEmailPreference = {
  __typename?: 'JourneysEmailPreference';
  accountNotifications: Scalars['Boolean']['output'];
  email: Scalars['String']['output'];
  unsubscribeAll: Scalars['Boolean']['output'];
};
// ...
export type Language = {
  __typename?: 'Language';
  audioPreview?: Maybe<AudioPreview>;
  bcp47?: Maybe<Scalars['String']['output']>;
  countries: Array<Country>;
  id: Scalars['ID']['output'];
  iso3?: Maybe<Scalars['String']['output']>;
  name: Array<LanguageName>;
};


export type LanguageNameArgs = {
  languageId?: InputMaybe<Scalars['ID']['input']>;
  primary?: InputMaybe<Scalars['Boolean']['input']>;
};

export enum LanguageIdType {
  Bcp47 = 'bcp47',
  DatabaseId = 'databaseId'
}

export type LanguageName = {
  __typename?: 'LanguageName';
  language: Language;
  primary: Scalars['Boolean']['output'];
  value: Scalars['String']['output'];
};
// ...
export type PlausibleStatsAggregateResponse = {
  __typename?: 'PlausibleStatsAggregateResponse';
  /** Bounce rate percentage. */
  bounceRate?: Maybe<PlausibleStatsAggregateValue>;
  /**
   * The percentage of visitors who completed the goal. Requires an `event:goal`
   * filter or `event:goal` property in the breakdown endpoint
   */
  conversionRate?: Maybe<PlausibleStatsAggregateValue>;
  /**
   * The number of events (pageviews + custom events). When filtering by a goal,
   *  this metric corresponds to "Total Conversions" in the dashboard.
   */
  events?: Maybe<PlausibleStatsAggregateValue>;
  /** The number of pageview events. */
  pageviews?: Maybe<PlausibleStatsAggregateValue>;
  /**
   * The average time users spend on viewing a single page. Requires an
   * `event:page` filter or `event:page` property in the breakdown endpoint.
   */
  timeOnPage?: Maybe<PlausibleStatsAggregateValue>;
  /**
   * The number of pageviews divided by the number of visits.
   * Returns a floating point number. Currently only supported in Aggregate and
   * Timeseries endpoints.
   */
  viewsPerVisit?: Maybe<PlausibleStatsAggregateValue>;
  /** Visit duration in seconds. */
  visitDuration?: Maybe<PlausibleStatsAggregateValue>;
  /** The number of unique visitors. */
  visitors?: Maybe<PlausibleStatsAggregateValue>;
  /** The number of visits/sessions. */
  visits?: Maybe<PlausibleStatsAggregateValue>;
};
// ...
export type PowerBiEmbed = {
  __typename?: 'PowerBiEmbed';
  /** The embed token */
  accessToken: Scalars['String']['output'];
  /** The embed URL of the report */
  embedUrl: Scalars['String']['output'];
  /** The date and time (UTC) of token expiration */
  expiration: Scalars['String']['output'];
  /** The report ID */
  reportId: Scalars['String']['output'];
  /** The name of the report */
  reportName: Scalars['String']['output'];
};

export type Query = {
  __typename?: 'Query';
  adminJourney: Journey;
  /**
   * returns all journeys that match the provided filters
   * If no team id is provided and template is not true then only returns journeys
   * where the user is not a member of a team but is an editor or owner of the
   * journey
   */
  adminJourneys: Array<Journey>;
  adminJourneysReport?: Maybe<PowerBiEmbed>;
  block: Block;
  blocks: Array<Block>;
  countries: Array<Country>;
  country?: Maybe<Country>;
  customDomain: CustomDomain;
  customDomains: Array<CustomDomain>;
  getJourneyProfile?: Maybe<JourneyProfile>;
  getMyCloudflareImages?: Maybe<Array<Maybe<CloudflareImage>>>;
  getMyCloudflareVideo?: Maybe<CloudflareVideo>;
  getMyCloudflareVideos?: Maybe<Array<Maybe<CloudflareVideo>>>;
  getUserRole?: Maybe<UserRole>;
  hosts: Array<Host>;
  integrations: Array<Integration>;
  journey: Journey;
  journeyCollection: JourneyCollection;
  journeyCollections: Array<Maybe<JourneyCollection>>;
  /** Get a JourneyVisitor count by JourneyVisitorFilter */
  journeyVisitorCount: Scalars['Int']['output'];
  /** Get a list of Visitor Information by Journey */
  journeyVisitorsConnection: JourneyVisitorsConnection;
  journeys: Array<Journey>;
  journeysEmailPreference?: Maybe<JourneysEmailPreference>;
  journeysPlausibleStatsAggregate: PlausibleStatsAggregateResponse;
  /**
   * This endpoint allows you to break down your stats by some property.
   * If you are familiar with SQL family databases, this endpoint corresponds to
   * running `GROUP BY` on a certain property in your stats, then ordering by the
   * count.
   * Check out the [properties](https://plausible.io/docs/stats-api#properties)
   * section for a reference of all the properties you can use in this query.
   * This endpoint can be used to fetch data for `Top sources`, `Top pages`,
   * `Top countries` and similar reports.
   * Currently, it is only possible to break down on one property at a time.
   * Using a list of properties with one query is not supported. So if you want
   * a breakdown by both `event:page` and `visit:source` for example, you would
   * have to make multiple queries (break down on one property and filter on
   * another) and then manually/programmatically group the results together in one
   * report. This also applies for breaking down by time periods. To get a daily
   * breakdown for every page, you would have to break down on `event:page` and
   * make multiple queries for each date.
   */
  journeysPlausibleStatsBreakdown: Array<PlausibleStatsResponse>;
  journeysPlausibleStatsRealtimeVisitors: Scalars['Int']['output'];
  /**
   * This endpoint provides timeseries data over a certain time period.
   * If you are familiar with the Plausible dashboard, this endpoint
   * corresponds to the main visitor graph.
   */
  journeysPlausibleStatsTimeseries: Array<PlausibleStatsResponse>;
  language?: Maybe<Language>;
  languages: Array<Language>;
  listUnsplashCollectionPhotos: Array<UnsplashPhoto>;
  me?: Maybe<User>;
  searchUnsplashPhotos: UnsplashQueryResponse;
  tags: Array<Tag>;
  team: Team;
  teams: Array<Team>;
  user?: Maybe<User>;
  userByEmail?: Maybe<User>;
  userInvites?: Maybe<Array<UserInvite>>;
  userTeam: UserTeam;
  userTeamInvites: Array<UserTeamInvite>;
  userTeams: Array<UserTeam>;
  video?: Maybe<Video>;
  videos: Array<Video>;
  /** Get a single visitor */
  visitor: Visitor;
  /** A list of visitors that are connected with a specific team. */
  visitorsConnection: VisitorsConnection;
};
// ...
export type UnsplashPhoto = {
  __typename?: 'UnsplashPhoto';
  alt_description?: Maybe<Scalars['String']['output']>;
  blur_hash: Scalars['String']['output'];
  categories?: Maybe<Array<Scalars['String']['output']>>;
  color?: Maybe<Scalars['String']['output']>;
  created_at: Scalars['String']['output'];
  current_user_collections: Array<Scalars['String']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  height: Scalars['Int']['output'];
  id: Scalars['String']['output'];
  liked_by_user: Scalars['Boolean']['output'];
  likes: Scalars['Int']['output'];
  links: UnsplashPhotoLinks;
  updated_at?: Maybe<Scalars['String']['output']>;
  urls: UnsplashPhotoUrls;
  user: UnsplashUser;
  width: Scalars['Int']['output'];
};
// ...
export type UserRole = {
  __typename?: 'UserRole';
  id: Scalars['ID']['output'];
  roles?: Maybe<Array<Role>>;
  userId: Scalars['ID']['output'];
};
// ...
export type GetLanguageQuery = { __typename?: 'Query', languages: Array<{ __typename?: 'Language', id: string, name: Array<{ __typename?: 'LanguageName', value: string, primary: boolean }>, countries: Array<{ __typename?: 'Country', continent: Array<{ __typename?: 'CountryContinent', value: string, primary: boolean }> }> }> };